Modify files structure
