+++ RelativeLineNumbers.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RelativeLineNumbers", "RelativeLineNumbers.csproj", "{3C8D2698-5F6A-4FE0-B8D2-5712C8BFF34E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RelativeLineNumbers", "RelativeLineNumbers\RelativeLineNumbers.csproj", "{3C8D2698-5F6A-4FE0-B8D2-5712C8BFF34E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
