+++ RelativeLineNumbers/source.extension.vsixmanifest
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
+  <Metadata>
+    <Identity Id="RelativeLineNumbers.Zoltan.Klinger.ecf8ccdf-17d2-4c5a-9131-bb6887c7b6c3" Version="1.1.2" Language="en-US" Publisher="Zoltan Klinger" />
+    <DisplayName>RelativeLineNumbers</DisplayName>
+    <Description xml:space="preserve">Displays line numbers relative to the current cursor position. Useful for ViEmu and VsVim extension users.</Description>
+    <MoreInfo>https://github.com/zklinger/RelativeLineNumbers</MoreInfo>
+    <License>LICENSE.txt</License>
+    <Icon>rln.png</Icon>
+    <PreviewImage>rln_preview.png</PreviewImage>
+  </Metadata>
+  <Installation>
+    <InstallationTarget Id="Microsoft.VisualStudio.Community" Version="[11.0,16.0)" />
+    <InstallationTarget Version="[11.0,16.0)" Id="Microsoft.VisualStudio.IntegratedShell" />
+    <InstallationTarget Version="[11.0,16.0)" Id="Microsoft.VisualStudio.Pro" />
+    <InstallationTarget Version="[11.0,16.0)" Id="Microsoft.VisualStudio.Enterprise" />
+  </Installation>
+  <Dependencies>
+    <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.5,)" />
+  </Dependencies>
+  <Assets>
+    <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%|" />
+  </Assets>
+  <Prerequisites>
+    <Prerequisite Id="Microsoft.VisualStudio.Component.CoreEditor" Version="[15.0,16.0)" DisplayName="Visual Studio core editor" />
+  </Prerequisites>
+</PackageManifest>
